+++ src/wbs/console/supervisor/SupervisorAdditionOperandSpec.java
@@ -2,7 +2,6 @@ package wbs.console.supervisor;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-
 import wbs.console.module.ConsoleModuleData;
 import wbs.framework.application.annotations.PrototypeComponent;
 import wbs.framework.data.annotations.DataAttribute;
@@ -21,10 +20,10 @@ class SupervisorAdditionOperandSpec {
 	SupervisorAdditionStatsResolverSpec supervisorAdditionStatsResolverSpec;
 
 	@DataAttribute
-	Integer coefficient = 1;
+	Long coefficient = 1l;
 
 	@DataAttribute
-	Integer value = 1;
+	Long value = 1l;
 
 	@DataAttribute (
 		name = "resolver")
