+++ src/wbs/console/supervisor/SupervisorIntegerInConditionSpec.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-
 import wbs.console.module.ConsoleModuleData;
 import wbs.framework.application.annotations.PrototypeComponent;
 import wbs.framework.data.annotations.DataAttribute;
@@ -31,6 +30,6 @@ class SupervisorIntegerInConditionSpec {
 		direct = true,
 		childElement = "item",
 		valueAttribute = "value")
-	List<Integer> values;
+	List <Long> values;
 
 }
