+++ src/wbs/console/supervisor/SupervisorMultiplicationOperandSpec.java
@@ -2,7 +2,6 @@ package wbs.console.supervisor;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-
 import wbs.console.module.ConsoleModuleData;
 import wbs.framework.application.annotations.PrototypeComponent;
 import wbs.framework.data.annotations.DataAttribute;
@@ -22,10 +21,10 @@ class SupervisorMultiplicationOperandSpec {
 		supervisorMultiplicationStatsResolverSpec;
 
 	@DataAttribute
-	Integer power = 1;
+	Long power = 1l;
 
 	@DataAttribute
-	Integer value = 1;
+	Long value = 1l;
 
 	@DataAttribute (
 		name = "resolver")
