+++ src/wbs/integrations/smsarena/model/SmsArenaDlrReportLogSearch.java
@@ -2,12 +2,12 @@ package wbs.integrations.smsarena.model;
 
 import java.io.Serializable;
 
+import org.joda.time.Instant;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import lombok.experimental.Accessors;
-
-import org.joda.time.Instant;
 
 @Accessors (chain = true)
 @Data
@@ -20,6 +20,6 @@ class SmsArenaDlrReportLogSearch
 	Instant timestampAfter;
 	Instant timestampBefore;
 
-	Integer routeId;
+	Long routeId;
 
 }
