+++ src/wbs/smsapps/subscription/model/SubscriptionNumberSearch.java
@@ -2,10 +2,10 @@ package wbs.smsapps.subscription.model;
 
 import java.io.Serializable;
 
+import org.joda.time.Instant;
+
 import lombok.Data;
 import lombok.experimental.Accessors;
-
-import org.joda.time.Instant;
 
 @Accessors (fluent = true)
 @Data
@@ -13,7 +13,7 @@ public
 class SubscriptionNumberSearch
 	implements Serializable {
 
-	Integer subscriptionId;
+	Long subscriptionId;
 
 	String numberLike;
